remove helper file
